WIP: basic processing; no transformations yet
